refactor: move errors below internal/errors
